Replace switch(class(x), ...) statements with more robust alternatives
Conditional statements that compare 'class(x)' to a string literal or
to a set of string literals (e.g. 'if (class(x) == "Foo") ...', or
'if (class(x) %in% c("Foo", "Bar")) ...', or 'switch(class(x),
"Foo"=..., "Bar"=...)' are generally unreliable and fragile. Statements
that take inheritance into account (e.g. 'if (is(x, "Foo")) ...') should
almost always be used instead.